requirement.txt
pyjwt
pyperclip
